requirements.txt
streamlit
python-dotenv
pinecone-client
langchain
langchain-pinecone
langchain-community
langchain-text-splitters
groq
gTTS
pypdf
openai
tiktoken